# infrastructure/api-gateway/openapi.yaml
openapi: "3.0.1"
info:
  title: "Panda CRM API"
  description: "API Gateway for Panda CRM microservices"
  version: "1.0.0"
servers:
  - url: "https://api.pandacrm.com"
    description: "Production"
  - url: "https://api-staging.pandacrm.com"
    description: "Staging"
...
paths:
...
  /api/contacts:
    get:
      summary: List contacts
      tags: [Contacts]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/contacts"
        type: http_proxy
        httpMethod: GET
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
    post:
      summary: Create contact
      tags: [Contacts]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/contacts"
        type: http_proxy
        httpMethod: POST
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
  /api/contacts/{id}:
    get:
      summary: Get contact by ID
      tags: [Contacts]
      security:
        - CognitoAuthorizer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/contacts/{id}"
        type: http_proxy
        httpMethod: GET
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
        requestParameters:
          integration.request.path.id: method.request.path.id
    put:
      summary: Update contact
      tags: [Contacts]
      security:
        - CognitoAuthorizer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/contacts/{id}"
        type: http_proxy
        httpMethod: PUT
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
        requestParameters:
          integration.request.path.id: method.request.path.id
    delete:
      summary: Delete contact
      tags: [Contacts]
      security:
        - CognitoAuthorizer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/contacts/{id}"
        type: http_proxy
        httpMethod: DELETE
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
        requestParameters:
          integration.request.path.id: method.request.path.id

  # Leads Service
...
  /api/payments:
    get:
      summary: List payments
      tags: [Payments]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/payments"
        type: http_proxy
        httpMethod: GET
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
    post:
      summary: Create payment
      tags: [Payments]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/payments"
        type: http_proxy
        httpMethod: POST
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
  /api/payments/{id}:
    get:
      summary: Get payment by ID
      tags: [Payments]
      security:
        - CognitoAuthorizer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/payments/{id}"
        type: http_proxy
        httpMethod: GET
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
        requestParameters:
          integration.request.path.id: method.request.path.id
...
components:
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - "arn:aws:cognito-idp:us-east-2:679128292059:userpool/us-east-2_e02zbxuZ2"